file structure/folder refactor
