+++ src/components/Homepage/Slider/SliderContent.jsx
