+++ src/components/Homepage/Slider/slider.css
