+++ src/pages/Homepage/index.jsx
@@ -1,6 +1,6 @@
-import Slider from "../../components/Slider";
+import Slider from "../../components/Homepage/Slider";
 import ProductGrid from "../../components/ReusableComponents/ProductGrid";
-import CategoryGrid from "../../components/CategoryGrid";
+import CategoryGrid from "../../components/Homepage/CategoryGrid";
 
 function Homepage() {
     return (
